requirements.txt
# tensorflow==2.17
opencv-python
numpy
pandas
filetype
fastapi
ultralytics==8.3.33
uvicorn
python-multipart
python-dotenv
pydantic-settings
pydantic
minio
transformers==4.49.0
pillow==10.4.0